NvVarsFileLib: Fix build warning with VS2005 on X64.
git-svn-id: https://edk2.svn.sourceforge.net/svnroot/edk2/trunk/edk2@9305 6f19259b-4bc3-4df7-8a09-765794883524

+++ OvmfPkg/Library/NvVarsFileLib/FsAccess.c
@@ -507,11 +507,11 @@ SaveNvVarsToFs (
     Status = PackVariableIntoFile (
                File,
                VariableName,
-               VariableNameSize,
+               (UINT32) VariableNameSize,
                &VendorGuid,
                VariableAttributes,
                VariableData,
-               VariableDataSize
+               (UINT32) VariableDataSize
                );
     if (EFI_ERROR (Status)) {
       break;
